requirements.txt
pandas
numpy
xgboost
scikit-learn
matplotlib
seaborn
imbalanced-learn
pymysql
mysql-connector-python
flask
